code refactoring move components in new folder
