+++ src/main.js
@@ -24,3 +24,4 @@ import './modules/modal';
 import './modules/baseLayer';
 import './modules/layerManager';
 import './modules/notifications';
+import './modules/examples';
