+++ src/modules/commons/index.js
@@ -5,4 +5,3 @@ import './components/checkbox';
 import './components/icon';
 import './components/autocomplete';
 import './components/spinner';
-import './components/examples';
